Add @export annotations to observer methods.
The @export annotations were not really designed for private methods, but this is a stop-gap patch for #91 which seems
to be tickling some bug in either this code or the closure compiler. This fixes a problem where the observer methods get
minimized but the definitions inside the `observers` property do not get minimized causing js errors when it cannot find
the observer methods.