Reset `tail` as routes deactivate.
This change proposes that `tail` should be reset when a route
deactivates. This has two important side-effects that need to be
considered:

 1. Child routes that no longer match the current path will become
 "inactive". Note that this is already true today in some circumstances,
 but not others, so the change brings consistency across all types of
 route path changes.
 2. The `tail` values down a routing branch will blink on and off in
 ways that they did not before. This is important to consider for
 performance purposes.

Regarding #2, the justification for this change is that `data` will
still remain consistent despite any thrashing of the `tail`, so UI
state can rely on `data` to avoid thrashing.